chore: export component definition update logic
