+++ cmd/init/definitionupdater/updater_test.go
@@ -1,4 +1,4 @@
-package main
+package definitionupdater
 
 import (
 	"testing"
